[spi] Formatting code with new code style

+++ thirdeye-spi/src/main/java/org/apache/pinot/thirdeye/dataframe/BooleanSeries.java
@@ -284,7 +284,7 @@ public final class BooleanSeries extends TypedSeries<BooleanSeries> {
     if (isNull(value)) {
       return DoubleSeries.NULL;
     }
-    return (double) value;
+    return value;
   }
 
   @Override
